// js_modules/ui-core/src/instance/backfill/types/useResumeBackfill.types.ts
/** Internal type. DO NOT USE DIRECTLY. */
type Exact<T extends {[key: string]: unknown}> = {[K in keyof T]: T[K]};
/** Internal type. DO NOT USE DIRECTLY. */
export type Incremental<T> =
  | T
  | {[P in keyof T]?: P extends ' $fragmentName' | '__typename' ? T[P] : never};
// Generated GraphQL types, do not edit manually.

import * as Types from '../../../graphql/types';

export type ResumeBackfillMutationVariables = Exact<{
  backfillId: string;
}>;

export type ResumeBackfillMutation = {
  __typename: 'Mutation';
  resumePartitionBackfill:
    | {
        __typename: 'PythonError';
        message: string;
        stack: Array<string>;
        errorChain: Array<{
          __typename: 'ErrorChainLink';
          isExplicitLink: boolean;
          error: {__typename: 'PythonError'; message: string; stack: Array<string>};
        }>;
      }
    | {__typename: 'ResumeBackfillSuccess'; backfillId: string}
    | {__typename: 'UnauthorizedError'; message: string};
};